+++ src/main/java/fr/uvsq/poo/monprojet/Joueur.java
@@ -1,33 +1,24 @@
-package fr.uvsq.poo.monprojet;
+package src.main.Projet;
 import java.util.*;
 
 
 public class Joueur {
 	int x,y;
 	String dir;
-	int vie;
-	String type;
 	int puissance;
-	ArrayList<Objet> inventaire;
 	
 	/* 	
-	 *	créer un joueur de type PNJ ou PJ en fonction de l'argument
-	 *	positionné aléatoirement dans une matrice salle
+	 *	crï¿½er un joueur de type PNJ ou PJ en fonction de l'argument
+	 *	positionnï¿½ alï¿½atoirement dans une matrice salle
 	 *	vie de base = 3, puissance de base = 30, si PNJ pas de vie
 	 *
 	 */
 	
-	public Joueur(String type) {
+	public Joueur() {
 		x = (int) (Math.random() * (Variables.largeur_salle - 1)) + 1;
 		y = (int) (Math.random() * (Variables.hauteur_salle - 1)) + 1;
 		dir = "haut";
-		this.type = type;
-		this.puissance = 30;
-		if(this.type == "PJ") {
-			this.vie = 3;
-			this.inventaire = new ArrayList<Objet>();
-		}
-		else this.vie = 0;		
+		this.puissance = 0;
 	}
 	
 	public int getPosX() {
@@ -52,9 +43,5 @@ public class Joueur {
 		return this;
 	}
 	
-	public ArrayList<Objet> PickUp(Objet Obj){
-		 this.inventaire.add(Obj);
-		 return this.inventaire;
-	}
 	
 }
