+++ src/main/java/fr/uvsq/poo/monprojet/Pnj.java
@@ -1,0 +1,15 @@
+
+public class Pnj extends Joueur {
+	
+
+	public Pnj() {
+		super();
+	}
+	
+	public Joueur Move(String dir) {
+		super.Move(dir);
+		return this;
+		
+	}
+
+}
